feat: add new bss-core package
