+++ .changeset/wet-singers-dream.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/batch-submitter-service': patch
+---
+
+Restructure to use bss-core package
