+++ go/bss-core/.gitignore
@@ -1,0 +1,1 @@
+/batch-submitter
