+++ go/bss-core/sentry_log.go
@@ -1,4 +1,4 @@
-package batchsubmitter
+package bsscore
 
 import (
 	"errors"
